refactor: rename grabit to clipit, to avoid pypi naming collisions
