+++ RELEASE.md
@@ -4,7 +4,7 @@
 - [ ] Sync the content in README.md with the latest changes in the app and git log. Manually review them as they'll most likely contain some gibberish as well.
 
 ###
-@grabit.py @readme.md
+@clipit.py @readme.md
 
 Help me update the readme file with the latest updates in the app. Remember to split the release notes into new features and fixes. If it helps, here's the git log:
 
@@ -33,4 +33,4 @@ git push origin main --tags
 
 
 - [ ] Create a new release on GitHub with the release notes
-- [ ] Make sure to include the latest grabit.py in the release
+- [ ] Make sure to include the latest clipit.py in the release
