+++ src/clipit/__init__.py
@@ -1,10 +1,10 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from grabit_md.core import GrabitError, OutputFormat
-from grabit_md.grabber import Grabber
+from clipit.core import ClipitError, OutputFormat
+from clipit.grabber import Grabber
 
 try:
-    __version__ = version("grabit-md")
+    __version__ = version("clipit")
 except PackageNotFoundError:
     # Fallback when running from source without installation
     __version__ = "dev"
@@ -13,5 +13,5 @@ __all__ = [
     "__version__",
     "Grabber",
     "OutputFormat",
-    "GrabitError",
+    "ClipitError",
 ]
