+++ src/clipit/core/downloader.py
@@ -1,7 +1,7 @@
 import requests
 from requests import RequestException
 
-from grabit_md.core import GrabitError
+from clipit.core import ClipitError
 
 
 def download_html_content(url, user_agent: str | None) -> str:
@@ -18,6 +18,6 @@ def download_html_content(url, user_agent: str | None) -> str:
         response.raise_for_status()
         html_content = response.content.decode("utf-8")
     except RequestException as e:
-        raise GrabitError(f"Error downloading {url}: {e}")
+        raise ClipitError(f"Error downloading {url}: {e}")
 
     return html_content
