+++ src/clipit/core/extractor.py
@@ -1,6 +1,6 @@
 from readabilipy import simple_json_from_html_string
 
-from grabit_md import GrabitError
+from clipit import ClipitError
 
 
 def extract_readable_content_and_title(html_content, use_readability_js):
@@ -13,14 +13,14 @@ def extract_readable_content_and_title(html_content, use_readability_js):
             rpy = simple_json_from_html_string(html_content, use_readability=False)
             content_html = rpy.get("content", "")
             if not content_html:
-                raise GrabitError("No content found")
+                raise ClipitError("No content found")
 
         content_html = content_html.replace(
             'href="about:blank/', 'href="../'
         )  # Fix for readability replacing ".." with "about:blank"
         title = (rpy.get("title") or "").strip()
     except Exception as e:
-        raise GrabitError(f"Error processing HTML content: {e}")
+        raise ClipitError(f"Error processing HTML content: {e}")
     return (
         content_html,
         title,
