+++ src/clipit/core/misc.py
@@ -1,0 +1,2 @@
+class ClipitError(Exception):
+    pass
