+++ src/clipit/core/output_format.py
