+++ src/clipit/grabbers/__init__.py
@@ -1,0 +1,4 @@
+from clipit.grabbers.base_grabber import BaseGrabber
+from clipit.grabbers.reddit_grabber import RedditGrabber
+
+__all__ = ["BaseGrabber", "RedditGrabber"]
