--- src/grabit_md/core/misc.py
@@ -1,2 +1,0 @@
-class GrabitError(Exception):
-    pass
