--- src/grabit_md/grabbers/__init__.py
@@ -1,4 +1,0 @@
-from grabit_md.grabbers.base_grabber import BaseGrabber
-from grabit_md.grabbers.reddit_grabber import RedditGrabber
-
-__all__ = ["BaseGrabber", "RedditGrabber"]
