+++ tests/test_file_outputs.py
@@ -1,5 +1,6 @@
 import pytest
-from grabit_md.core.writer import sanitize_filename
+
+from clipit.core.writer import sanitize_filename
 
 
 @pytest.mark.parametrize(
