+++ tests/test_markdown_conversion.py
@@ -1,4 +1,4 @@
-from grabit_md.core.markdown_converter import convert_to_markdown
+from clipit.core.markdown_converter import convert_to_markdown
 
 
 def test_converts_bold_and_italics_with_different_markup():
